add treasure model

+++ TreasureDungeons/Source/Models/Cube.swift
@@ -49,7 +49,7 @@ class Cube : Model {
         
     ]
     
-    let indexList : [GLubyte] = [
+    let indexList : [GLuint] = [
         
         // Front
         0, 1, 2,
