+++ TreasureDungeons/Source/Models/Wall.swift
@@ -49,7 +49,7 @@ class Wall : Model {
         
     ]
     
-    let indexList : [GLubyte] = [
+    let indexList : [GLuint] = [
         
         // Front
         0, 1, 2,
